+++ app/models/warehouse_receipt.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, String, Date, Boolean, Float, ForeignKey
+from sqlalchemy import Column, String, Date, Boolean,  ForeignKey, Integer
 from .base import Base
 from sqlalchemy.dialects.postgresql import UUID
 from sqlalchemy.orm import relationship
@@ -11,9 +11,14 @@ class WarehouseReceipt(Base):
     estimated_date = Column(Date, unique = False, nullable = False)
     delivery_status = Column(String, unique = False, nullable = False)
     payment_status = Column(String, unique = False, nullable = False)
-    subtotal = Column(Float, unique = False, nullable = False)
-    promotion = Column(Float, unique = False, nullable = True)
-    total = Column(Float, unique = False, nullable = False)
+    subtotal = Column(Integer, unique = False, nullable = False)
+    promotion = Column(Integer, unique = False, nullable = True)
+    total = Column(Integer, unique = False, nullable = False)
+    # created_by = Column(UUID(as_uuid=True), unique = False, nullable = False)
+    # belong_to_vendor = Column(UUID(as_uuid=True), unique = False, nullable = False)
+    # belong_to_contract = Column(UUID(as_uuid=True), unique = False, nullable = False)
+    # belong_to_invoice = Column(UUID(as_uuid=True), unique = False, nullable = True)
+    
     created_by = Column(UUID(as_uuid=True), ForeignKey('employee.id'), unique = False, nullable = False)
     belong_to_vendor = Column(UUID(as_uuid=True), ForeignKey('vendor.id'), unique = False, nullable = False)
     belong_to_contract = Column(UUID(as_uuid=True), ForeignKey('contract.id'), unique = False, nullable = False)
