+++ app/schemas/contract_for_product.py
@@ -4,13 +4,13 @@ from pydantic import BaseModel, UUID4
 class ContractForProductCreateParams(BaseModel):
     contract_id: UUID4
     product_id: UUID4
-    price: float
+    price: int
     
 class ContractForProductCreate(BaseModel):
     id: UUID4
     contract_id: UUID4
     product_id: UUID4
-    price: float
+    price: int
 
 class ContractForProductUpdate(BaseModel):
     pass
